XCM configs
